chore: Update dependencies and styles
This commit updates the dependencies in the `package.json` file, including the `next` version to `14.2.5` and adds the `react-icons` package as a dependency. It also modifies the global CSS file `globals.css` to remove unused styles and adds a box-sizing reset to all elements. Additionally, the layout component `layout.tsx` is updated to include a new header component and a footer component.